requirements.txt

openai==1.11.0
pyyaml==6.0.1
